HOTT-2913 Refactored finding date for descendants
* Takes into account start/end dates
* Refactored code to further reduce duplication
