+++ app/lib/changes_table_populator/commodity_code_end_dated.rb
@@ -14,17 +14,6 @@ module ChangesTablePopulator
       { validity_end_date: (previous_day.beginning_of_day..previous_day.end_of_day) }
     end
 
-    def build_all_change_records(source_changes)
-      source_changes.map do |source_change|
-        last_valid_day = (day - 1.day).beginning_of_day
-        declarable_on_last_day = end_line?(row: source_change, day: last_valid_day)
-
-        build_change_record(row: source_change,
-                            day:,
-                            is_end_line: declarable_on_last_day)
-      end
-    end
-
     def change_type
       'commodity'
     end
