+++ app/lib/changes_table_populator/commodity_code_started.rb
@@ -12,14 +12,6 @@ module ChangesTablePopulator
       { validity_start_date: day }
     end
 
-    def build_all_change_records(source_changes)
-      source_changes.map do |source_change|
-        build_change_record(row: source_change,
-                            day:,
-                            is_end_line: end_line?(row: source_change, day:))
-      end
-    end
-
     def change_type
       'commodity'
     end
