+++ app/lib/changes_table_populator/descendant_populator.rb
@@ -1,0 +1,17 @@
+module ChangesTablePopulator
+  module DescendantPopulator
+    def build_all_change_records(source_changes)
+      source_changes
+        .uniq { |element| element[:goods_nomenclature_sid] }
+        .collect_concat do |source_change|
+          build_descendant_change_records(row: source_change, day:)
+        end
+    end
+
+    def build_descendant_change_records(row:, day: Time.zone.today)
+      find_source_and_descendants(row:, day:).map do |descendant|
+        build_change_record(row: descendant, day:, is_end_line: descendant.ns_declarable?)
+      end
+    end
+  end
+end
