+++ app/lib/changes_table_populator/measure_created_or_updated.rb
@@ -1,5 +1,7 @@
 module ChangesTablePopulator
   class MeasureCreatedOrUpdated < Importer
+    include DescendantPopulator
+
     def source_table
       :measures
     end
@@ -13,14 +15,6 @@ module ChangesTablePopulator
                  '(validity_end_date IS NULL OR validity_end_date > ?) AND ' \
                  'operation IN (\'C\', \'U\') AND ' \
                  'operation_date = ?', day, day, day)
-    end
-
-    def build_all_change_records(source_changes)
-      source_changes
-        .uniq { |element| element[:goods_nomenclature_sid] }
-        .collect_concat do |source_change|
-          build_descendant_change_records(row: source_change, day:)
-        end
     end
 
     def change_type
