+++ app/lib/changes_table_populator/measure_end_dated.rb
@@ -1,5 +1,7 @@
 module ChangesTablePopulator
   class MeasureEndDated < Importer
+    include DescendantPopulator
+
     def source_table
       :measures
     end
@@ -12,14 +14,6 @@ module ChangesTablePopulator
       previous_day = (day - 1.day)
 
       { validity_end_date: (previous_day.beginning_of_day..previous_day.end_of_day) }
-    end
-
-    def build_all_change_records(source_changes)
-      source_changes
-        .uniq { |element| element[:goods_nomenclature_sid] }
-        .collect_concat do |source_change|
-          build_descendant_change_records(row: source_change, day:)
-        end
     end
 
     def change_type
